+++ packages/BiostructureViewer/CHANGELOG.md
@@ -42,8 +42,10 @@
 * Fix pdbqt parser to get autodock-gpu scores
 * Fix PdbGridCellRenderer, NglGlDocService to not mix cells
 * Fix MolstarViewer to handle ligandColumnName changed
-* BsV: Add test MolstarViewer for pdb_data with cell Click
-* BsV: Add test MolstarViewer for pdb_id, data provider
+* Add test MolstarViewer for pdb_data with cell Click
+* Add test MolstarViewer for pdb_id, data provider
+* Fix biostructure data provider to return JSON string
+* Add test biostructure provider with binary data (bcif)
 
 ## 1.0.11 (2023-07-24)